Moves dependencies to store module
